Move AnnotateEntity() to ActivityService
